requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118
timm==1.0.15
pillow==11.0.0
numpy==1.26.4
tqdm==4.67.1
scikit-learn==1.6.1
transformers==4.30.1
six==1.17.0
pyyaml==6.0.2
matplotlib==3.9.4
pandas==2.2.3
jieba==0.42.1
opencv-python==4.11.0.86
ultralytics==8.3.116